// frontend/src/api/generated/models/authResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Clinic API
 * OpenAPI spec version: v1
 */

export interface AuthResponse {
  /** @nullable */
  accessToken?: string | null;
  /** @nullable */
  refreshToken?: string | null;
  accessTokenExpiresAt?: string;
  userId?: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  role?: string | null;
}